Spin off states.h
